refactor: 파일 경로 정리
